+++ app/src/main/java/com/gratus/meditationtrakcer/ReleaseNotesActivity.java
@@ -9,6 +9,8 @@ import androidx.core.view.WindowInsetsCompat;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.gratus.meditationtrakcer.adapters.ReleaseNotesAdapter;
+
 import java.io.BufferedReader;
 import java.io.InputStream;
 import java.io.InputStreamReader;