+++ app/src/main/java/com/gratus/meditationtrakcer/ReportsActivity.java
@@ -6,15 +6,18 @@ import android.widget.ImageButton;
 import android.widget.LinearLayout;
 
 import androidx.activity.EdgeToEdge;
+import androidx.activity.OnBackPressedCallback;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.gratus.meditationtrakcer.dialogfragments.YMPickerDialogFragment;
 
 import com.google.android.material.button.MaterialButtonToggleGroup;
 
 public class ReportsActivity extends BaseActivity {
 
-    private ImageButton initiateButton;
+    private ImageButton btn_YearReport, btn_MonthReport;
     private LinearLayout buttonContainer, initiateButtonLayout, MYreport;
 
     @Override
@@ -30,12 +33,51 @@ public class ReportsActivity extends BaseActivity {
         // Initialize the toolbar and menu button
         setupToolbar(R.id.toolbar2, R.id.menubutton);
 
-    }
+        btn_YearReport = findViewById(R.id.btn_reportYear);
+        btn_MonthReport = findViewById(R.id.btn_reportMonth);
 
-    @Override
-    public void onBackPressed() {
-            // Default back behavior
-            super.onBackPressed();
-        }
+
+        btn_MonthReport.setOnClickListener(v -> {
+            YMPickerDialogFragment.newInstance(
+                    YMPickerDialogFragment.PickerMode.YEAR_MONTH,
+                    new YMPickerDialogFragment.OnYMSelectedListener() {
+                        @Override
+                        public void onYearSelected(int year) {}
+
+                        @Override
+                        public void onYearMonthSelected(int year, int month) {
+                            // handle month report
+                        }
+                    }
+            ).show(getSupportFragmentManager(), "ym_picker");
+        });
+
+        btn_YearReport.setOnClickListener(v -> {
+            YMPickerDialogFragment.newInstance(
+                    YMPickerDialogFragment.PickerMode.YEAR_ONLY,
+                    new YMPickerDialogFragment.OnYMSelectedListener() {
+                        @Override
+                        public void onYearSelected(int year) {
+                            // handle year report
+                        }
+
+                        @Override
+                        public void onYearMonthSelected(int year, int month) {}
+                    }
+            ).show(getSupportFragmentManager(), "y_picker");
+        });
+
+
+        getOnBackPressedDispatcher().addCallback(this, new OnBackPressedCallback(true) {
+            @Override
+            public void handleOnBackPressed() {
+                // Your custom logic here
+                // To simply go back:
+                setEnabled(false); // Disable this callback
+                getOnBackPressedDispatcher().onBackPressed(); // Call again
+            }
+        });
+
+    }
 }