+++ app/src/main/java/com/gratus/meditationtrakcer/YMPickerDialogFragment.java
@@ -1,6 +1,0 @@
-package com.gratus.meditationtrakcer;
-
-import androidx.fragment.app.DialogFragment;
-
-public class YMPickerDialogFragment extends DialogFragment {
-}