+++ app/src/main/java/com/gratus/meditationtrakcer/databasehelpers/StreakDatabaseHelper.java
@@ -1,4 +1,4 @@
-package com.gratus.meditationtrakcer;
+package com.gratus.meditationtrakcer.databasehelpers;
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -6,6 +6,8 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 import android.util.Log;
+
+import com.gratus.meditationtrakcer.datamodels.Streak;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;