+++ app/src/main/java/com/gratus/meditationtrakcer/datamodels/Goal.java
@@ -1,4 +1,4 @@
-package com.gratus.meditationtrakcer;
+package com.gratus.meditationtrakcer.datamodels;
 
 public class Goal {
     private int id;