+++ app/src/main/java/com/gratus/meditationtrakcer/datamodels/Streak.java
@@ -1,4 +1,4 @@
-package com.gratus.meditationtrakcer;
+package com.gratus.meditationtrakcer.datamodels;
 
 public class Streak {
     private int id;